+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/tinfoilsh/confidential-audio-processing/proxy
+
+go 1.23.5
