--- proxy/go.mod
@@ -1,3 +1,0 @@
-module github.com/tinfoilsh/confidential-audio-processing/proxy
-
-go 1.23.5
